requirements.txt
python-telegram-bot
requests
feedparser
transformers
torch
yfinance
